requirements.txt
# Core
pandas>=2.1
numpy>=1.26
yfinance>=0.2.40
lxml>=5.0
python-dotenv>=1.0

# Brokerage
alpaca-py>=0.30

# AI swarm
anthropic>=0.40

# Charting
matplotlib>=3.8

# Storage
pyarrow>=14.0

# Notifications
requests>=2.31

# Dev
pytest>=8.0
